split the settings-backed timezone helpers out of lib
lib/timezone.js imported getSettings from services so three of its exports could
resolve the user's configured timezone. That inverted the layering for a module
almost every scheduler depends on.

The split follows what each function actually needs, not where it happened to
live: everything that takes a timezone as an ARGUMENT is pure and stays in
lib/timezone.js (getLocalParts, getUtcOffsetMs, nextLocalTime, todayInTimezone,
parseHHMM, isWithinTimeWindow, the HHMM regexes). The three that READ it from
settings move to the new services/userTimezone.js: getUserTimezone,
getTimezoneUpdatedAt, userLocalToday.

lib/timezone.js now imports nothing from services.

userLocalToday keeps its original rationale, which the move preserves rather
than breaks: it exists so the scored-session and training-log services share one
day boundary without importing each other. A standalone third module satisfies
that as well as lib did.

No re-export shim here — a shim would have to point upward from lib into
services, which is the inversion being fixed. Consumers are re-pointed instead:
44 import sites (27 of which also pull pure helpers and are split across both
modules) plus 15 vi.mock factories and two in-test dynamic imports. The mocks
matter as much as the imports — a factory left aimed at lib/timezone.js would
silently stop applying and let the real settings read through.

Refs #4901.
