+++ server/lib/timezone.js
@@ -6,8 +6,6 @@
  * These helpers convert between UTC and the user's local timezone.
  */
 
-import { getSettings } from '../services/settings.js'
-
 const WEEKDAY_MAP = { Sun: 0, Mon: 1, Tue: 2, Wed: 3, Thu: 4, Fri: 5, Sat: 6 }
 
 // Cache Intl.DateTimeFormat instances per timezone — these are expensive to construct
@@ -30,43 +28,6 @@ function getFormatter(timezone) {
     formatterCache.set(timezone, fmt)
   }
   return fmt
-}
-
-/**
- * Get the user's configured timezone, falling back to system timezone.
- * @returns {Promise<string>} IANA timezone string (e.g., 'America/Los_Angeles')
- */
-export async function getUserTimezone() {
-  const settings = await getSettings()
-  const tz = settings.timezone
-  if (tz) {
-    // Validate the configured timezone; fall back to system timezone if invalid
-    try {
-      new Intl.DateTimeFormat('en-US', { timeZone: tz })
-      return tz
-    } catch {
-      console.error(`❌ Invalid configured timezone "${tz}", falling back to system default`)
-    }
-  }
-  return Intl.DateTimeFormat().resolvedOptions().timeZone
-}
-
-/**
- * Get the UTC-ms timestamp at which the user's `timezone` setting last actually
- * changed, or null when it has never been recorded. Stamped by settings.js's
- * write path (`save()`) only when the `timezone` value genuinely differs from
- * the prior on-disk value. Timezone-dependent schedulers use this to avoid
- * replaying a slot that "occurred" under a timezone that was no longer active —
- * see catchUpMissedSlot in meatspacePostReminder.js (#2040).
- *
- * Sentinel discipline: null (not a number, ≤ 0) means "never changed / unset",
- * so callers must NOT treat absence as a floor — an unset value gates nothing.
- * @returns {Promise<number|null>} UTC ms, or null when unset.
- */
-export async function getTimezoneUpdatedAt() {
-  const settings = await getSettings()
-  const ts = settings.timezoneUpdatedAt
-  return typeof ts === 'number' && ts > 0 ? ts : null
 }
 
 /**
@@ -146,27 +107,6 @@ export function nextLocalTime(afterMs, hours, minutes, timezone) {
 export function todayInTimezone(timezone, atDate = new Date()) {
   const parts = getLocalParts(atDate, timezone)
   return `${parts.year}-${String(parts.month).padStart(2, '0')}-${String(parts.day).padStart(2, '0')}`
-}
-
-/**
- * Today's `YYYY-MM-DD` in the USER's configured timezone — the resolve-tz-then-
- * derive-today combination that every per-day feature keying off the local day
- * needs (POST daily status/streaks in meatspacePost.js & meatspacePostTraining.js,
- * the Daily Driver). Kept here (in lib) so both the scored-session service and the
- * training-log service share ONE day boundary without importing each other (they
- * already have a mutual dependency; this avoids adding to it). The process runs
- * `TZ=UTC`, so deriving the day from a bare `new Date().toISOString()` would use
- * the server's UTC day and misfile activity around the local/UTC midnight boundary
- * for non-UTC users (issue #2681).
- * Pass `atDate` (the instant a writer already captured for its timestamp) so the
- * day key derives from the SAME instant, never a fresh `new Date()` sampled after
- * the awaited settings read — which could cross midnight and split the two fields
- * onto different days (issue #2681 r5).
- * @param {Date} [atDate] - instant to key (defaults to now)
- * @returns {Promise<string>}
- */
-export async function userLocalToday(atDate = new Date()) {
-  return todayInTimezone(await getUserTimezone(), atDate)
 }
 
 // ---------------------------------------------------------------------------
