+++ server/services/brainJournal.test.js
@@ -20,9 +20,11 @@ vi.mock('../lib/fileUtils.js', async () => {
 });
 
 vi.mock('../lib/timezone.js', () => ({
-tryReadFile: vi.fn().mockResolvedValue(null),
-  getUserTimezone: () => Promise.resolve('UTC'),
+  tryReadFile: vi.fn().mockResolvedValue(null),
   todayInTimezone: () => '2026-04-17',
+}));
+vi.mock('./userTimezone.js', () => ({
+  getUserTimezone: () => Promise.resolve('UTC'),
 }));
 
 vi.mock('./obsidian.js', () => ({
