+++ server/services/whatsappImport.js
@@ -53,7 +53,7 @@ import { readFile } from 'fs/promises';
 import { createHash } from 'crypto';
 import { collectZipEntries, isZipUpload } from '../lib/zipStream.js';
 import { shortSummary, recordEvents, resolveEventInstant } from './humanActivity.js';
-import { getUserTimezone } from '../lib/timezone.js';
+import { getUserTimezone } from './userTimezone.js';
 
 // ---------------------------------------------------------------------------
 // Pure helpers (exported for unit tests — no DB, no filesystem, no side effects).
