+++ server/services/youtubeSync.js
@@ -30,7 +30,8 @@
  * exported and unit-tested against saved DOM-shaped fixtures — no browser or DB.
  */
 import { dataPath, ensureDir, atomicWrite, tryReadFile, safeJSONParse, sleep } from '../lib/fileUtils.js';
-import { getUserTimezone, todayInTimezone } from '../lib/timezone.js';
+import { todayInTimezone } from '../lib/timezone.js';
+import { getUserTimezone } from './userTimezone.js';
 import { getSettings } from './settings.js';
 import { findOrOpenPage, listCdpPages, isAuthPage, evaluateOnPage } from './browserService.js';
 import { localDayRangeUtc, shortSummary } from './humanActivity.js';
